First totally working tree
