+++ src/gordion/tree.py
@@ -44,6 +44,23 @@ class Tree(gordion.Repository):
     return listed_path
 
   def ensure(self, url: str = ''):
+    # Derive url if necessary.
+    if not url:
+      assert gordion.Repository._exists(self.path)
+      self.handle = git.Repo(self.path)
+      self.url = self.handle.remotes.origin.url
+    else:
+      self.url = url
+      if gordion.Repository._exists(self.path):
+        if self.url != self.handle.remotes.origin.url:
+          self._check_different_repo_same_path(self._root())
+          gordion.Tree._safe_remove_repo(self.path)
+
+    # Check for different repositories cloned to the same path.
+    self._check_different_repo_same_path(self._root())
+
+    # Check for a duplicates repository cloned at different paths.
+    self._check_duplicate_repo_path(self._root())
     super().ensure(url)
     self.yeditor.reload()
 
@@ -146,3 +163,37 @@ class Tree(gordion.Repository):
         child.ensure(child_info['url'])
         child.update(child_info['tag'], branch_name, force)
         self.children[child_name] = child
+
+  def _check_different_repo_same_path(self, other):
+    """
+    Recursively checks the repository path against another repository and it's children.
+    """
+    host, username, repo_name = gordion.extract_repo_details(self.url)
+    other_host, other_username, other_repo_name = gordion.extract_repo_details(other.url)
+
+    # Check if the remote repository is the same
+    if host != other_host or username != other_username or repo_name != other_repo_name:
+      # Make sure the repository does not have the same local path.
+      if self.path == other.path:
+        raise gordion.UpdateDifferentRepoSamePathError(self, other)
+
+    # Check against the other's children
+    for _, other_child in other.children.items():
+      Tree._check_different_repo_same_path(self, other_child)
+
+  def _check_duplicate_repo_path(self, other):
+    """
+    Recursively checks the repository path against another repository and it's children.
+    """
+    host, username, repo_name = gordion.extract_repo_details(self.url)
+    other_host, other_username, other_repo_name = gordion.extract_repo_details(other.url)
+
+    # Check if the remote repository is the same
+    if host == other_host and username == other_username and repo_name == other_repo_name:
+      # Make sure the repository has the same local path.
+      if self.path != other.path:
+        raise gordion.UpdateDuplicateRepoPathError(self, other)
+
+    # Check against the other's children
+    for _, other_child in other.children.items():
+      Tree._check_duplicate_repo_path(self, other_child)
